requirements.txt
pymongo==4.6.1
flask==3.0.0
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-limiter==3.5.0
bcrypt==4.1.2
python-dotenv==1.0.0
user-agents==2.2.0